// charting_library/broker-api.d.ts
export declare const enum OrderType {
	Limit = 1,
	Market = 2,
	Stop = 3,
	StopLimit = 4
}
export declare const enum ParentType {
	Order = 1,
	Position = 2,
	Trade = 3
}
export declare const enum PriceType {
	Limit = 1,
	Stop = 2
}
export declare const enum Side {
	Buy = 1,
	Sell = -1
}
// ...
export declare enum StopType {
	StopLoss = 0,
	TrailingStop = 1
}
// ...
export interface BracketOrder extends BracketOrderBase, CustomFields {
}
export interface BracketOrderBase extends PlacedOrderBase {
	/** If order is a bracket then this should contain base order/position id. */
	parentId: string;
	/** Type of the bracket's parent */
	parentType: ParentType;
}
// ...
export interface CustomFields {
	/** Custom field */
	[key: string]: any;
}
// ...
export interface CustomInputFieldsValues {
	[fieldId: string]: TextWithCheckboxValue | boolean | string | any;
}
// ...
export interface OrderDuration {
	/**
	 * type is OrderDurationMetaInfo.value
	 */
	type: string;
	/** Order duration time */
	datetime?: number;
}
// ...
export interface OrderTemplate {
	/** Symbol identifier */
	symbol: string;
	/** Order Type */
	type?: OrderType;
	/** order / execution side */
	side?: Side;
	/** Order quantity */
	qty?: number;
	/** Type of Stop Order */
	stopType?: StopType;
	/** Order stop price */
	stopPrice?: number;
	/** Order limit price */
	limitPrice?: number;
	/** Order Take Profit (Brackets) */
	takeProfit?: number;
	/** Order Stop loss (Brackets) */
	stopLoss?: number;
	/** Order Trailing stop (Brackets) */
	trailingStopPips?: number;
	/** Duration or expiration of an order */
	duration?: OrderDuration;
	/** Custom input fields */
	customFields?: CustomInputFieldsValues;
}
export interface PipValues {
	/** value of 1 pip if you buy */
	buyPipValue: number;
	/** value of 1 pip if you sell */
	sellPipValue: number;
}
export interface PlaceOrderResult {
	/** Order id */
	orderId?: string;
}
export interface PlacedOrder extends PlacedOrderBase, CustomFields {
}
/**
 * Info about a placed order
 */
export interface PlacedOrderBase {
	/** Order ID */
	id: string;
	/** Symbol name */
	symbol: string;
	/** Order type */
	type: OrderType;
	/** Order side (buy or sell) */
	side: Side;
	/** Order quantity (double) */
	qty: number;
	/** Order status */
	status: OrderStatus;
	/** Stop loss price (double). Available when Brackets are enabled */
	stopLoss?: number;
	/** Trailing stop Pips value (double). Available when Brackets are enabled */
	trailingStopPips?: number;
	/** Stop Loss type. It should be set to 1 (StopType.TrailingStop) for trailing stop orders. */
	stopType?: StopType;
	/** Take profit price (double). Available when Brackets are enabled */
	takeProfit?: number;
	/** Order duration */
	duration?: OrderDuration;
	/**
	 * An object that contains the results of broker specific user inputs (for example a digital signature).
	 * There are two possible kinds of custom fields: an input field with a checkbox and a custom combobox.
	 */
	customFields?: CustomInputFieldsValues;
	/** Filled order quantity (double) */
	filledQty?: number;
	/** Average fulfilled price for the order (double) */
	avgPrice?: number;
	/** Last update time (unix timestamp in milliseconds) */
	updateTime?: number;
	/** Price for the limit order (double) */
	limitPrice?: number;
	/** Price for the stop order (double) */
	stopPrice?: number;
	/** Message describing the state of the order */
	message?: OrderOrPositionMessage;
}
// ...
export interface PreOrder extends OrderTemplate {
	/** @inheritDoc */
	symbol: string;
	/** @inheritDoc */
	type: OrderType;
	/** @inheritDoc */
	side: Side;
	/** @inheritDoc */
	qty: number;
	/** Current Quotes */
	currentQuotes?: AskBid;
	/**
	 * It is set to `true`, if the order closes a position.
	 */
	isClose?: boolean;
}
// ...
export interface TextWithCheckboxValue {
	/** Checkbox text */
	text: string;
	/** Whether the checkbox is checked */
	checked: boolean;
}
// ...
export interface TradingQuotes {
	/** Trade */
	trade?: number;
	/** Quote size */
	size?: number;
	/** Bid price */
	bid?: number;
	/** Bid quantity */
	bid_size?: number;
	/** Ask price */
	ask?: number;
	/** Ask quantity */
	ask_size?: number;
	/** Price spread */
	spread?: number;
	/** Whether quotes are delayed */
	isDelayed?: boolean;
	/** Whether quotes are halted */
	isHalted?: boolean;
	/** Whether quotes are hard to borrow */
	isHardToBorrow?: boolean;
	/** Whether quotes are can not be shorted */
	isNotShortable?: boolean;
}
// ...
export type AskBid = Required<Pick<TradingQuotes, "ask" | "bid">>;
// ...
export type Order = PlacedOrder | BracketOrder;